+++ projetobeauty/src/main/java/br/com/beautyfit/exceptions/KeywordNotFoundException.java
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 public class KeywordNotFoundException extends RuntimeException {
 	private static final long serialVersionUID = 1L;
 
-	public KeywordNotFoundException(Long id, String message) {
+	public KeywordNotFoundException(Integer id, String message) {
 		super(message + id);
 	}
 }
