+++ projetobeauty/src/main/java/br/com/beautyfit/repository/CustomerServiceRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.CrudRepository;
 import br.com.beautyfit.model.CustomerService;
 
 @Transactional
-public interface CustomerServiceRepository  extends CrudRepository<CustomerService, Long> {
+public interface CustomerServiceRepository  extends CrudRepository<CustomerService, Integer> {
 	
 	public List<CustomerService> findAll();
